feat: moved RedisStreamConsumer package and deleted unuseless classes
